Create static constructor for PartitionFileReferenceStats
